chore: remove old auditlog persister
